# PolyclinicLab/Polyclinic.Client.Wasm/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Polyclinic API",
    "version": "v1"
  },
  "paths": {
    "/api/patients": {
      "get": {
        "tags": ["Patients"],
        "operationId": "GetPatients",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Patients"],
        "operationId": "CreatePatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdatePatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/patients/{id}": {
      "get": {
        "tags": ["Patients"],
        "operationId": "GetPatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Patients"],
        "operationId": "UpdatePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdatePatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Patients"],
        "operationId": "DeletePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/doctors": {
      "get": {
        "tags": ["Doctors"],
        "operationId": "GetDoctors",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Doctors"],
        "operationId": "CreateDoctor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdateDoctorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctors/{id}": {
      "get": {
        "tags": ["Doctors"],
        "operationId": "GetDoctor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Doctors"],
        "operationId": "UpdateDoctor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdateDoctorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Doctors"],
        "operationId": "DeleteDoctor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
...
  },
  "components": {
    "schemas": {
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "passport": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "bloodType": {
            "type": "string",
            "nullable": true
          },
          "rhFactor": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateUpdatePatientDto": {
        "type": "object",
        "properties": {
          "passport": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "bloodType": {
            "type": "string",
            "nullable": true
          },
          "rhFactor": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "DoctorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "passport": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "birthYear": {
            "type": "integer",
            "format": "int32"
          },
          "specialization": {
            "type": "string",
            "nullable": true
          },
          "experience": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CreateUpdateDoctorDto": {
        "type": "object",
        "properties": {
          "passport": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "birthYear": {
            "type": "integer",
            "format": "int32"
          },
          "specialization": {
            "type": "string",
            "nullable": true
          },
          "experience": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}